% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_dsb_d.R, R/fill_dsb_d.R
\name{color_dsb_d}
\alias{color_dsb_d}
\alias{colour_dsb_d}
\title{Apply a discrete color scale to your \code{ggplot2} graphics}
\usage{
color_dsb_d(
  palette = "main",
  primary = NULL,
  secondary = NULL,
  reverse = FALSE,
  ...
)

colour_dsb_d(
  palette = "main",
  primary = NULL,
  secondary = NULL,
  reverse = FALSE,
  ...
)

color_dsb_d(
  palette = "main",
  primary = NULL,
  secondary = NULL,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{a continuous color palette. Default is "main"}

\item{primary}{a name of a valid DSB color (see \link[epinionDSB]{dsb_colvec}) to be applied to the primary category. Invalid with more than three levels}

\item{secondary}{a name of a valid DSB color (see \link[epinionDSB]{dsb_colvec}) to be applied to the primary category. Invalid with more than three levels}

\item{reverse}{a logical value indicating whether to reverse the color scale. Defaults to \code{FALSE}}
}
\description{
This function applies a discrete color scale using colors from the DSB color palette

This function applies a discrete color scale using colors from the DSB color palette
}
\examples{
library(tidyverse)
ggplot(mtcars, aes(x = wt, y = mpg, color = factor(am))) +
geom_point() +
color_dsb_d()
library(tidyverse)
ggplot(mtcars, aes(x = wt, y = mpg, factor(am))) +
geom_point() +
colour_dsb_d()
library(tidyverse)
ggplot(mtcars, aes(x = wt, y = mpg, fill = factor(am))) +
geom_point(shape = 21, color = "NA") +
fill_dsb_d()
}
\author{
Jeppe Vierø

Jeppe Vierø
}
