% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_epi_c.R
\name{color_epi_c}
\alias{color_epi_c}
\title{Use an Epinion styled color palette for your ggplot2 plots}
\usage{
color_epi_c(palette = "epiblue", reverse = FALSE, ...)
}
\arguments{
\item{...}{x}
}
\value{
... z
}
\description{
This function creates a DSB compliant color palette for continuous variables
}
\author{
Jeppe Vierø
}
