% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_epi_d.R
\name{color_epi_d}
\alias{color_epi_d}
\title{Use an Epinion styled discrete color palette for your ggplot2 plots}
\usage{
color_epi_d(
  palette = "epinion",
  primary = "Epi_DarkBlue",
  other = "Epi_WarmSand",
  reverse = FALSE,
  ...
)
}
\arguments{
\item{...}{x}
}
\value{
... z
}
\description{
This function creates a DSB compliant color palette for discrete variables
}
\author{
Jeppe Vierø
}
