% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_dsb.R
\name{colour_dsb}
\alias{colour_dsb}
\title{Use a DSB styled color palette for your ggplot2 plots}
\usage{
colour_dsb(palette = "two", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{...}{x}
}
\value{
... z
}
\description{
This function creates a DSB compliant color palette
}
\author{
Jeppe Vierø
}
