% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_dsb_c.R
\name{colour_dsb_c}
\alias{colour_dsb_c}
\title{Use a DSB styled color palette for your ggplot2 plots}
\usage{
colour_dsb_c(palette = "reds", reverse = FALSE, ...)
}
\arguments{
\item{...}{x}
}
\value{
... z
}
\description{
This function creates a DSB compliant color palette for continuous variables
}
\author{
Jeppe Vierø
}
