% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsb_colors.R
\docType{data}
\name{dsb_colors}
\alias{dsb_colors}
\title{Extract HEX-codes and RGB-codes by color name}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 3 columns.
}
\usage{
dsb_colors
}
\description{
This snippet defines a dataset of colors in the DSB Design Manual and provides their corresponding color codes in two formats:
RGB and HEX.
The code provided here is for transparency purposes only, the resulting data is provided in the data folder (dsb_colors.rda)
}
\author{
Jeppe Vierø
}
\keyword{datasets}
