% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsb_colvec.R
\name{dsb_colvec}
\alias{dsb_colvec}
\title{Use a DSB styled color palette for your ggplot2 graphics}
\usage{
dsb_colvec(...)
}
\arguments{
\item{...}{takes as input the name of the color(s) needed}
}
\value{
the corresponding HEX code
}
\description{
This function extract DSB compliant colors (in HEX codes) by their name
}
\author{
Jeppe Vierø
}
