% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsb_grabcol.R
\name{dsb_grabcol}
\alias{dsb_grabcol}
\title{Extract colors (HEX) from the DSB color palette by their name}
\usage{
dsb_grabcol(color)
}
\arguments{
\item{color}{a DSB color (see \code{names(epinionDSB::dsb_cols)}) with or without Epinion prefix}
}
\value{
the corresponding HEX code
}
\description{
This function extracts colors (in HEX codes) from the DSB color palette by their official name
}
\author{
Jeppe Vierø
}
