% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsb_pal.R
\name{dsb_pal}
\alias{dsb_pal}
\title{Use a DSB styled continuous color palette for your ggplot2 plots}
\usage{
dsb_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{...}{x}
}
\value{
... z
}
\description{
This function creates a DSB compliant color palette
}
\author{
Jeppe Vierø
}
