% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsb_pal_d.R
\name{dsb_pal_d}
\alias{dsb_pal_d}
\title{Use a DSB styled discrete color palette for your ggplot2 plots}
\usage{
dsb_pal_d(
  palette = "main",
  primary = "DSB_Red",
  other = "DSB_DarkBlue",
  reverse = FALSE
)
}
\arguments{
\item{...}{x}
}
\value{
... z
}
\description{
This function creates a DSB compliant color palette
}
\author{
Jeppe Vierø
}
