% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsb_style.R
\name{dsb_style}
\alias{dsb_style}
\title{Add a DSB style to your graph}
\usage{
dsb_style(legend = TRUE)
}
\arguments{
\item{legend}{a logical value indicating whether a legend is included. Defaults to TRUE.}
}
\value{
A CVI compliant plot
}
\description{
This function adds a DSB compliant theme to ggplot figures. It can be added at the end of your ggplot syntax.
}
\author{
Jeppe Vierø
}
