% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsbpal_d}
\alias{dsbpal_d}
\title{Discrete DSB color palettes}
\format{
A \code{list} with 1 element
}
\usage{
dsbpal_d
}
\description{
A dataset containing the location of 10,000 US cities
}
\keyword{datasets}
