% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_colvec.R
\name{epi_colvec}
\alias{epi_colvec}
\title{Use an Epinion styled color palette for your ggplot2 graphics}
\usage{
epi_colvec(...)
}
\arguments{
\item{...}{takes as input the name of the color(s) needed}
}
\value{
the corresponding HEX code
}
\description{
This function extract Epinion compliant colors (in HEX codes) by their name
}
\author{
Jeppe Vierø
}
