% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_pal.R
\name{epi_pal}
\alias{epi_pal}
\title{Use an Epinion styled continuous color palette for your ggplot2 plots}
\usage{
epi_pal(palette = "epinion", reverse = FALSE, ...)
}
\arguments{
\item{...}{x}
}
\value{
... z
}
\description{
This function creates a DSB compliant color palette
}
\author{
Jeppe Vierø
}
