% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_pal_d.R
\name{epi_pal_d}
\alias{epi_pal_d}
\title{Use an Epinion styled discrete color palette for your ggplot2 plots}
\usage{
epi_pal_d(
  palette = "epinion",
  primary = "Epi_DarkBlue",
  other = "Epi_WarmSand",
  reverse = FALSE
)
}
\arguments{
\item{...}{x}
}
\value{
... z
}
\description{
This function creates a DSB compliant color palette
}
\author{
Jeppe Vierø
}
