% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_dsb.R
\name{scale_fill_dsb}
\alias{scale_fill_dsb}
\title{Use a DSB styled color palette for your ggplot2 plots}
\usage{
scale_fill_dsb(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{...}{x}
}
\value{
... z
}
\description{
This function creates a DSB compliant color palette
}
\author{
Jeppe Vierø
}
