% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_dsb_d.R
\name{color_dsb_d}
\alias{color_dsb_d}
\title{Use a DSB styled discrete color palette for your ggplot2 plots}
\usage{
color_dsb_d(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{...}{x}
}
\value{
... z
}
\description{
This function creates a DSB compliant color palette for discrete variables
}
\author{
Jeppe Vierø
}
