% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_dsb_d.R
\name{colour_dsb_d}
\alias{colour_dsb_d}
\title{Use a DSB styled color palette for your ggplot2 plots}
\usage{
colour_dsb_d(
  ...,
  values,
  aesthetics = "colour",
  breaks = waiver(),
  na.value = "grey50"
)
}
\arguments{
\item{...}{x}
}
\value{
... z
}
\description{
This function creates a DSB compliant color palette for discrete variables
}
\author{
Jeppe Vierø
}
