% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_dsb_d.R
\name{colour_dsb_d}
\alias{colour_dsb_d}
\title{Use a DSB styled discrete color palette for your ggplot2 plots}
\usage{
colour_dsb_d(
  palette = "main",
  primary = "DSB_Red",
  other = "DSB_DarkBlue",
  reverse = FALSE,
  ...
)
}
\arguments{
\item{...}{x}
}
\value{
... z
}
\description{
This function creates a DSB compliant color palette for discrete variables
}
\author{
Jeppe Vierø
}
