% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsb_colvec.R
\name{dsb_colvec}
\alias{dsb_colvec}
\title{Extract colors from the Epinion CVI color palette by their name}
\usage{
dsb_colvec(...)
}
\arguments{
\item{...}{takes as input the name of the color(s) needed}
}
\value{
the corresponding HEX code
}
\description{
This function extracts colors (in HEX codes) from the DSB color palette by their name
}
\author{
Jeppe Vierø
}
