% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dsb.R
\name{fill_dsb_d}
\alias{fill_dsb_d}
\title{Use a DSB styled color palette for your ggplot2 plots}
\usage{
fill_dsb_d(palette = "two", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{...}{x}
}
\value{
... z
}
\description{
This function creates a DSB compliant color palette
}
\author{
Jeppe Vierø
}
